license: Add license documents and specify version
The readme says this project is licensed under the LGPL, but does
not specify a version or include the license text. This commit adds
the LGPL license as COPYING.LESSER and the GPL which it requires
as COPYING. These are the v3+ versions of the license.

This commit also applies formatting fixes to README.md